Refactor file locations, sort by timestamp
