+++ assets/src/components/activities/directed-discussion/discussion/discussion-hook.tsx
